Update .NET Standard 2.1, enable nullable context
Update the project to reference the latest .NET Standard 2.1 and enable nullable reference types for the project. This helps better communicate deisgn intent especially when declaring fields of reference types such as List<>.

Also, add VB in addition to C# in the nuget package description, since this is a .NET Standard library so it can be used in all sorts or projects in a language agnostic manner.